move get sharing slide util
