+++ functions/src/lib/exercise.ts
@@ -1,7 +1,4 @@
-import {
-  Exercise,
-  ExerciseSlideSharingSlide,
-} from '../../../shared/src/types/generated/Exercise';
+import {Exercise} from '../../../shared/src/types/generated/Exercise';
 import i18next, {DEFAULT_LANGUAGE_TAG, LANGUAGE_TAG} from './i18n';
 
 export const getExerciseById = (
@@ -19,15 +16,4 @@ export const getExerciseById = (
   }
 
   return;
-};
-
-export const getSharingSlideById = (
-  exercise: Exercise | undefined,
-  sharingId: string,
-) => {
-  return exercise?.slides
-    .filter(s => s.type === 'sharing')
-    .find(s => (s as ExerciseSlideSharingSlide)?.id === sharingId) as
-    | ExerciseSlideSharingSlide
-    | undefined;
 };
